trying to fix size_t

+++ printf/num_format.c
@@ -27,7 +27,7 @@ unsigned long long int		unsigned_len(va_list *ap, t_flag arg_flag)
 	else if (arg_flag.e_length == j)
 		num = (intmax_t)va_arg(*ap, unsigned long long int);
 	else if (arg_flag.e_length == ll)
-		num = (ssize_t)va_arg(*ap, unsigned long long int);
+		num = (size_t)va_arg(*ap, unsigned long long int);
 	else
 		num = (unsigned int)va_arg(*ap, unsigned long long int);
 	return (num);
@@ -48,7 +48,7 @@ long long int				signed_len(va_list *ap, t_flag arg_flag)
 	else if (arg_flag.e_length == j)
 		num = (intmax_t)va_arg(*ap, long long int);
 	else if (arg_flag.e_length == z)
-		num = (ssize_t)va_arg(*ap, long long int);
+		num = va_arg(*ap, long long int);
 	else
 		num = (int)va_arg(*ap, long long int);
 	return (num);
